Fix SpawnOnMap objects not scaling on zoom
The current implementation never sets a parent, so the references to local position and scale in the `Update` function just refer to global position and rotation. This change sets the `AbstractMap _map` as the parent of the newly instantiated object and sets local scale according against that. For position we set global instead of local position because `GeoToWorldPosition` returns the global position.