Use UniAndroid to check permission status
In the existing implementation, `_gotPermissionRequestResponse` is not reliably set when the user sets permission, which can leave the application in a broken state that can only be resolved by force-quitting and restarting. Using UniAndroid instead seems to fix the issue and results in the application being in a working state after the user accepts the permission request.